refactor: reorganize traits ext
